refactor: separation of concerns
- create data, core, and utils dirs
- move business logic from main.py to app.py
- move data (csv, sqlite db) to data dir
