+++ utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility modules for the Lunch application."""
